requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cloudinary==1.41.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
daphne==4.1.2
decorator==5.1.1
Django==4.2.5
django-cloudinary-storage==0.3.0
django-extensions==3.2.3
djangorestframework==3.14.0
executing==2.1.0
fonttools==4.55.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipython==8.23.0
jedi==0.19.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
numpy==1.26.4
oauthlib==3.2.2
packaging==24.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
pyOpenSSL==24.3.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
pywin32==308
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.2
stack-data==0.6.3
tornado==6.4
traitlets==5.14.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
zope.interface==7.2
